# improve GPU performance of transpose