<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] --> Performance optimization

### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] --> OPs

### Describe
<!-- Describe what this PR does --> Use 32BitIndex to improve GPU performance of transpose.

#### Performance
Test on V100, cuda 10
|op | input shape | input dtype | perm|before |after|speed up|
|---|---|---|---|---|---|---|
|transpose| [4, 512, 512, 1] |float32|[1, 2, 0, 3]|0.0171 ms|0.0133 ms  | 1.29x|
|transpose| [4, 512, 512, 1] |float16|[1, 2, 0, 3]|0.0139 ms|0.0065 ms  | 2.14x|
|transpose| [4, 12, 512, 64] |float32|[2, 0, 1, 3]|0.0248 ms|0.0192 ms  | 1.29x|
|transpose| [4, 12, 512, 64] |float16|[2, 0, 1, 3]|0.0208 ms|0.0099 ms  | 2.10x|